requirements.txt
Django==1.8
MySQL-python==1.2.5
Pillow==2.8.1
Whoosh==2.6.0
argparse==1.2.1
django-extensions==1.5.2
django-filters==0.1.0
django-haystack==2.3.1
six==1.9.0
wsgiref==0.1.2
